backend/lambda/python/requirements.txt
boto3==1.34.0
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.20
anthropic==0.8.1
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
pydantic==2.5.0
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.12.0
flake8==6.1.0